+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix cssnano: @font-family minify bugs
 * Fix projectName variables
 * Edit `Getting started` on `README.MD`
+* Add new-page function (read README)
 
 # 0.3
 * Clear faviconData.json file